+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/fiori-elements-writer
 
+## 3.1.16
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/ui5-test-writer 1.2.34 → 1.2.35
+
+## 3.1.15
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fe-fpm-writer 1.3.8 → 1.3.9
+
 ## 3.1.14
 
 ### Patch Changes